jclass: track location info to nested class/interface definitions
Pity it has to be placed so precisely: adding stx's location to the
begin form wouldn't work. To be fair though, if doing so, should the
cj-sourcify-deep take the next outer or inner location info?
(Currently it takes the toplevel location instead. Which also mostly
works since it's "the macro use location".)