cj-posix: replace all ISO-8859-1-string usage with UTF-8-string
This does solve compilation on clang++. It didn't accept conversion of
'unsigned char *' to 'const char *'.

What is the reason for the unsigned?

This will break handling of non-UTF-8 paths, right? (ISO would have
allowed to handle all paths, even if the Scheme string would not be
proper ('human readable').)