create dot-oo-optim.scm
Works solely with the type name, is this really always OK? Looks so
naive.

This runs about 6-7 times faster than the normal dispatch without even
using cj-standarddeclares.

(Given only ~10% of the cost of a normal dispatch is GC cost, only a
small part will be the avoidance of list allocation.)