statprof: add encoding to HTML output
Hard code it, just like the assumtion in all of the files in
chj-schemelib. Or more to the point, the assumed reader encoding
setting. Could retrieve that? But, actually it just has to match what
the encoding on the output port is actually set to.

It uses with-output-to-file without explicit setting, which is
arguably a bug. So to match it up with the encoding here, it has to
rely on my -:t8,f8,-8 settings (or Gambit's default if now the same,
don't remember).