rename {list,stream}-length>{,=} to add minus after "length"
For consistency with |length-=| and |length->=| from
`predicates-1.scm`.

It is tempting to go the other way and remove all minuses. "After all,
|string>?| doesn't have a minus either." Except, it does have a
question mark, inconsistent anyway, and the difference is that those
compare two of the same type. Whereas here it's about comparing the
length of one argument to a integer argument. There's even |length=|
in `predicates-1.scm` already.