safer-apply: do not optimize away variables
This appears to be the most difficult to debug case when issues happen
at macro expansion times. Nothing in backtrace except compiler
internals, thus seeing anything in ,e becomes the last hope.

Oh, doesn't change anything? cont was already there? But actually
(show-def cont) will lead to the macro definition in question, good.

And from the value in the exception I can deduct the call site of the
macro, too, why is that not shown right away?, aha, since safer-apply
doesn't raise a source-error.

Aha: source being passed to an object constructor was just incidental;
the constructor is of course not expecting to deal with source thus
giving a non-source-error exception.

Example (paths shortened):

  *** ERROR IN #<procedure #3 safer-apply>, "lib/cj-source-util--include.scm"@146.16 -- assignments does not match (source-of list?): [[source1] [[[source1] t "lib/clojure.scm" 721481] [[source1] ([[source1] :test "lib/clojure.scm" 918089] [[source1] v "lib/clojure.scm" 1311305]) "lib/clojure.scm" 852553]] "lib/clojure.scm" 655945]

  > ,b
  0  safer-apply    "lib/cj-source-util--include.scm"@146:16
  1  | _eval|
  2  ##comp-body
  3  ##comp-lambda-aux
  4  ##comp-top-define
  ... (long and nothing in my code comes up)

Aha |assignments| is in if-let.scm, yes, which is the macro for which
cont is in its expander, too.