if-let: use assert* not cj-typed for macros
This is a deficiency in cj-typed that I hope to fix soon-ish.

cj-typed should throw source-error in this case; or predicates should
return source error false values (and cj-typed then simply throw
those). Or exceptions should follow a protocol on where context/values
are and cj-exception-handler should use that (perhaps with CAN.) to
get the value and if it is a source value then show its location,
perhaps.