show,*: add |show|, wrapper around .show
So that it is possible to modify .show's behaviour (lexically
limited), i.e. create another generic, pass that generic down to .show
so as to create a modification of |show| (lexically, without global
monkey patching).

Soo.. this would actually be unnecessary with lexical (local) monkey
patching? How does it work, fully, please?